Fix row hexagon spacing
For a square area covered by a hexagonal grid, where the hexagons are in the pointy orientation (⬡) the number of hexagons in the X dimension is determined by the spacing.
However the number of rows in the Y dimension is not not equal to the number of tiles in X. The tiles in Y scale as: (x_spacing * sqrt(3)) / 2 to the x spacing.
I had not implemented this, so the matplotlib hexbin function was stretching the hexagons in the Y dimension. Now it is almost fixed, but there is still a very slight deviation (<2e-9%) probably a rounding error.